+++ src/main/java/chess/controller/GameController.java
@@ -6,8 +6,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+@RestController
 @RequestMapping("/games")
 public class GameController {
 
@@ -15,11 +16,6 @@ public class GameController {
 
     public GameController(GameService gameService) {
         this.gameService = gameService;
-    }
-
-    @GetMapping("/{id}")
-    public String getGameById() {
-        return "chess";
     }
 
     @GetMapping("/room")
